moved some generic classes to a module of their own (serialiser.py) and implemented a factory function to instantiate appropriate scheduler class
